fix bug

+++ QLTHUCHI/src/controller/LogsController.java
@@ -150,7 +150,7 @@ public class LogsController {
         LogsDB logsDB = new LogsDB();
         logsDB.updateData(logDataToTable(new String[]{"id", "idOfItemInCategory", "price", "note"}));
     }
-    public void updateDataRow(int id, String note, int amountS){
+    public void updateDataRow(int id, String note, double amountS){
         // int amount = Integer.parseInt("100");
         for (int i = 0; i < this.logs.getSize(); i++) {
             if(this.logs.getLogAt(i).getID() == id){
